Web programming-php-card storage
A website consisting of 10 web pages that I programmed during university. It allows the users to log in or register, after which they can buy, sell or examine their pokemon card collection.
This project was designed and programmed by me using html and php. The images used are available for public use on the official pokemon website. It was a good way to practice logins, input verification and general php usage.