Hobby project-html-javascript-videogame
This is a hobby project, which is under development. A simple game called Path of Destiny, which I designed in my freetime, and later started implementing in this format. It uses html css and a lot of javascript. The art was done with AI to substitute my lacking artistic talent, but everything else was done entirely by me, including game design, testing, implementation, balancing and more. 
Game programming is a small niche, but it requires proficiency with the language used, as some game concepts can be extremely complex. It was/is a good way to practice my coding skills.